+++ Pipfile
@@ -8,6 +8,7 @@ pylint = "*"
 isort = "*"
 
 [dev-packages]
+pytest = "*"
 
 [requires]
 python_version = "3.9"
